# backend/src/main/java/com/academia/students/dto/SportsProfileDto.java
package com.academia.students.dto;

import com.academia.students.DominantHand;
import com.academia.students.SportsProfileEntity;
import java.time.Instant;

/**
 * Bloque deportivo, vista de administrador: ficha completa ({@link StudentAdminDto}) y
 * respuesta de {@code PUT /students/{id}/sports-profile}. Incluye {@code coachNotes}; la
 * familia recibe {@link StudentGuardianDto.SportsProfile}, que no lo tiene.
 */
public record SportsProfileDto(
        String level,
        DominantHand dominantHand,
        String ranking,
        String previousClub,
        String history,
        String goals,
        String coachNotes,
        Instant updatedAt) {

    public static SportsProfileDto from(SportsProfileEntity entity) {
        return new SportsProfileDto(entity.getLevel(), entity.getDominantHand(), entity.getRanking(),
                entity.getPreviousClub(), entity.getHistory(), entity.getGoals(), entity.getCoachNotes(),
                entity.getUpdatedAt());
    }
}